refactor: move HTTP handlers out of cron_jobs into routes/cron_jobs
cron_jobs.rs is a domain module — it should not import Axum or know about
HTTP status codes. Extract the 7 route handlers into routes/cron_jobs.rs.

- New routes/cron_jobs.rs: create/list/get/update/replace/delete/trigger
  with their #[utoipa::path] decorators; imports axum, calls svc_* service fns
- cron_jobs.rs: remove all Axum imports (Path, State, StatusCode, Json)
  and all handler functions; now only domain types + service layer
- routes/mod.rs: add pub mod cron_jobs
- routes/http.rs: alias `use super::cron_jobs as jobs` and reference
  jobs::* in the Router instead of cron_jobs::*
- openapi.rs: update handler paths from crate::cron_jobs::* to
  crate::routes::cron_jobs::*; schema refs (CronJob etc.) unchanged
