+++ src/openapi.rs
@@ -16,13 +16,13 @@
         crate::routes::http::health,
         crate::routes::http::echo,
         crate::routes::http::list_system_cron_jobs,
-        crate::cron_jobs::list,
-        crate::cron_jobs::create,
-        crate::cron_jobs::get,
-        crate::cron_jobs::replace,
-        crate::cron_jobs::update,
-        crate::cron_jobs::delete,
-        crate::cron_jobs::trigger,
+        crate::routes::cron_jobs::list,
+        crate::routes::cron_jobs::create,
+        crate::routes::cron_jobs::get,
+        crate::routes::cron_jobs::replace,
+        crate::routes::cron_jobs::update,
+        crate::routes::cron_jobs::delete,
+        crate::routes::cron_jobs::trigger,
     ),
     components(schemas(
         crate::cron_jobs::CronJob,
